refactor(core): restructure Quarry command handling and update dependencies
- Remove package-lock.json to streamline dependency management.
- Update the stratal dependency in package.json to use the latest version.
- Refactor command imports in application.ts and quarry.ts for improved clarity.
- Introduce QuarryRegistry class to manage command registration and execution.
- Enhance command execution flow by deferring instantiation to call-time.
- Update tests to reflect changes in command wiring and registry structure.
