+++ examples/17-commands/package.json
@@ -8,7 +8,7 @@
   },
   "dependencies": {
     "reflect-metadata": "^0.2.2",
-    "stratal": "file:../../packages/core"
+    "stratal": "latest"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20251014.0",
