+++ packages/core/src/application.ts
@@ -14,13 +14,13 @@ import { ConsoleTransport, JsonFormatter, LOGGER_TOKENS, LoggerService, LogLevel
 import { ModuleRegistry } from './module/module-registry'
 import type { DynamicModule, ModuleClass } from './module/types'
 import { OpenAPIModule } from './openapi'
+import type { Command } from './quarry/command'
+import { QuarryRegistry } from './quarry/quarry-registry'
+import type { CommandInput, CommandResult } from './quarry/types'
 import { type ConsumerRegistry } from './queue/consumer-registry'
 import type { IQueueConsumer, QueueMessage } from './queue/queue-consumer'
 import { type QueueManager } from './queue/queue-manager'
 import { QueueModule } from './queue/queue.module'
-import type { Command } from './quarry/command'
-import { QuarryRegistry } from './quarry/quarry'
-import type { CommandInput, CommandResult } from './quarry/types'
 import { type IController, type RouterContext } from './router'
 import { HonoApp } from './router/hono-app'
 import type { VersioningOptions } from './router/types'
