+++ packages/core/src/quarry/index.ts
@@ -2,7 +2,7 @@ export { Command } from './command'
 export { CommandNotFoundError } from './errors/command-not-found.error'
 export { CommandError } from './errors/command.error'
 export { isCommand } from './is-command'
-export { QuarryRegistry } from './quarry'
+export { QuarryRegistry } from './quarry-registry'
 export { parseSignature } from './signature-parser'
 export type { CommandInput, CommandResult, ParsedArgument, ParsedOption, ParsedSignature, Quarry } from './types'
 export { generateUsage } from './usage-generator'
