+++ packages/core/src/quarry/quarry-registry.ts
